+++ reana_server/ext.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of REANA.
-# Copyright (C) 2019, 2020, 2021 CERN.
+# Copyright (C) 2019, 2020, 2021, 2022 CERN.
 #
 # REANA is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -10,10 +10,47 @@
 
 import logging
 
+from flask import jsonify
+from flask_limiter.errors import RateLimitExceeded
 from flask_menu import Menu
+from marshmallow.exceptions import ValidationError
 from reana_commons.config import REANA_LOG_FORMAT, REANA_LOG_LEVEL
+from werkzeug.exceptions import UnprocessableEntity
 
 from reana_server import config
+
+
+def handle_rate_limit_error(error: RateLimitExceeded):
+    """Error handler for flask_limiter exception ``RateLimitExceeded``.
+
+    This error handler is needed to display useful error messages, instead of the
+    generic default one, when rate limit exception is raised.
+    """
+    rate_limit = error.description or str(error)
+    error_message = (
+        f'Too many requests in a given amount of time. Only "{rate_limit}" allowed.'
+    )
+    return jsonify({"message": error_message}), 429
+
+
+def handle_args_validation_error(error: UnprocessableEntity):
+    """Error handler for werkzeug exception ``UnprocessableEntity``.
+
+    This error handler is needed to display useful error messages, instead of the
+    generic default one, when marshmallow argument validation fails.
+    """
+    error_message = error.description or str(error)
+
+    exception = getattr(error, "exc", None)
+    if isinstance(exception, ValidationError):
+        validation_messages = []
+        for field, messages in exception.normalized_messages().items():
+            validation_messages.append(
+                "Field '{}': {}".format(field, ", ".join(messages))
+            )
+        error_message = ". ".join(validation_messages)
+
+    return jsonify({"message": error_message}), 400
 
 
 class REANA(object):
@@ -31,6 +68,7 @@ class REANA(object):
     def init_app(self, app):
         """Flask application initialization."""
         self.init_config(app)
+        self.init_error_handlers(app)
         Menu(app=app)
 
         @app.teardown_appcontext
@@ -62,3 +100,8 @@ class REANA(object):
         for k in dir(config):
             if k.startswith("REANA_"):
                 app.config.setdefault(k, getattr(config, k))
+
+    def init_error_handlers(self, app):
+        """Initialize custom error handlers."""
+        app.register_error_handler(RateLimitExceeded, handle_rate_limit_error)
+        app.register_error_handler(UnprocessableEntity, handle_args_validation_error)